+++ src/main/java/com/example/ISAprojekat/Repository/CottageRepository.java
@@ -4,5 +4,5 @@ import com.example.ISAprojekat.Model.BoatOwner;
 import com.example.ISAprojekat.Model.Cottage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CottageRepository extends JpaRepository<Cottage,Long> {
+public interface CottageRepository extends JpaRepository<Cottage,Integer> {
 }
