+++ src/main/java/com/example/ISAprojekat/Service/AdminService.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface AdminService {
     Admin getByUsernameAndPassword(String username, String password);
      List<Admin> findAll();
-    Admin getOne(Long id);
-    Admin findOne(Long id);
+    Admin getOne(Integer id);
+    Admin findOne(Integer id);
     Admin update(Admin admin) throws Exception;
 }
