+++ src/main/java/com/example/ISAprojekat/Service/AdventureService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface AdventureService {
     List<Adventure> findAll();
     List<Adventure> findAllByAdventureName(String adventureName);
-    Adventure findOne(Long id);
+    Adventure findOne(Integer id);
     Adventure save(Adventure adventure) throws Exception;
-    void delete(Long id);
+    void delete(Integer id);
     Adventure update(Adventure adventure) throws Exception;
     List<Adventure> getAllAdventuresFromInstructor(String email);
 }
