+++ src/main/java/com/example/ISAprojekat/Service/CottageService.java
@@ -6,9 +6,9 @@ import com.example.ISAprojekat.Model.CottageOwner;
 import java.util.List;
 
 public interface CottageService {
-    Cottage getOne(Long id);
+    Cottage getOne(Integer id);
     List<Cottage> findAll();
     Cottage save (Cottage cottage) throws Exception;
-    void delete(Long id);
+    void delete(Integer id);
     Cottage update(Cottage cottage);
 }
