+++ src/main/java/com/example/ISAprojekat/Service/RegKorisnikService.java
@@ -3,5 +3,5 @@ package com.example.ISAprojekat.Service;
 import com.example.ISAprojekat.Model.RegKorisnik;
 
 public interface RegKorisnikService {
-    RegKorisnik getOne(Long id);
+    RegKorisnik getOne(Integer id);
 }
